Revert "move networkclustering module to clustering.spatial"
