+++ pypsa/__init__.py
@@ -6,7 +6,6 @@ Energy system modelling library.
 """
 
 from pypsa import (
-    clustering,
     components,
     contingency,
     descriptors,
