--- pypsa/clustering/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Clustering functionality for PyPSA networks.
-"""
-
-from pypsa.clustering import spatial, temporal
