--- pypsa/clustering/temporal.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Functions for temporal clustering of networks.
-"""
-
-__author__ = (
-    "PyPSA Developers, see https://pypsa.readthedocs.io/en/latest/developers.html"
-)
-__copyright__ = (
-    "Copyright 2023 PyPSA Developers, see https://pypsa.readthedocs.io/en/latest/developers.html, "
-    "MIT License"
-)
